+++ recipes/lora_dpo_distributed.py
@@ -582,7 +582,7 @@ class LoRADPORecipeDistributed(FTRecipeInterface):
         The core training loop.
         """
         # clean up before training begins
-        utils.cleanup_before_training()
+        training.cleanup_before_training()
 
         _, rank = utils.get_world_size_and_rank()
