+++ torchtune/training/__init__.py
@@ -5,7 +5,23 @@
 # LICENSE file in the root directory of this source tree.
 
 from torchtune.training.quantization import get_quantizer_mode
+from trochtune.training.memory import (
+    cleanup_before_training,
+    create_optim_in_bwd_wrapper,
+    get_memory_stats,
+    log_memory_stats,
+    OptimizerInBackwardWrapper,
+    register_optim_in_bwd_hooks,
+    set_activation_checkpointing,
+)
 
 __all__ = [
     "get_quantizer_mode",
+    "cleanup_before_training",
+    "create_optim_in_bwd_wrapper",
+    "get_memory_stats",
+    "log_memory_stats",
+    "OptimizerInBackwardWrapper",
+    "register_optim_in_bwd_hooks",
+    "set_activation_checkpointing",
 ]
